+++ homeassistant/components/mysensors/__init__.py
@@ -22,7 +22,7 @@ from .const import (
 from .device import get_mysensors_devices
 from .gateway import get_mysensors_gateway, setup_gateways, finish_setup
 
-REQUIREMENTS = ['pymysensors==0.17.0']
+REQUIREMENTS = ['pymysensors==0.18.0']
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -135,7 +135,7 @@ def setup_mysensors_platform(
     # Only act if called via MySensors by discovery event.
     # Otherwise gateway is not set up.
     if not discovery_info:
-        return
+        return None
     if device_args is None:
         device_args = ()
     new_devices = []
