+++ homeassistant/components/mysensors/const.py
@@ -16,10 +16,12 @@ CONF_TOPIC_OUT_PREFIX = 'topic_out_prefix'
 CONF_VERSION = 'version'
 
 DOMAIN = 'mysensors'
+MYSENSORS_GATEWAY_READY = 'mysensors_gateway_ready_{}'
 MYSENSORS_GATEWAYS = 'mysensors_gateways'
 PLATFORM = 'platform'
 SCHEMA = 'schema'
-SIGNAL_CALLBACK = 'mysensors_callback_{}_{}_{}_{}'
+CHILD_CALLBACK = 'mysensors_child_callback_{}_{}_{}_{}'
+NODE_CALLBACK = 'mysensors_node_callback_{}_{}'
 TYPE = 'type'
 
 # MySensors const schemas
