+++ homeassistant/components/mysensors/gateway.py
@@ -4,32 +4,28 @@ from collections import defaultdict
 import logging
 import socket
 import sys
-from timeit import default_timer as timer
 
 import async_timeout
 import voluptuous as vol
 
 from homeassistant.const import (
-    CONF_NAME, CONF_OPTIMISTIC, EVENT_HOMEASSISTANT_STOP)
+    CONF_OPTIMISTIC, EVENT_HOMEASSISTANT_STOP)
 from homeassistant.core import callback
-from homeassistant.helpers import discovery
 import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.setup import async_setup_component
 
 from .const import (
-    ATTR_DEVICES, CONF_BAUD_RATE, CONF_DEVICE, CONF_GATEWAYS, CONF_NODES,
+    CONF_BAUD_RATE, CONF_DEVICE, CONF_GATEWAYS, CONF_NODES,
     CONF_PERSISTENCE, CONF_PERSISTENCE_FILE, CONF_RETAIN, CONF_TCP_PORT,
     CONF_TOPIC_IN_PREFIX, CONF_TOPIC_OUT_PREFIX, CONF_VERSION, DOMAIN,
-    MYSENSORS_CONST_SCHEMA, MYSENSORS_GATEWAYS, PLATFORM, SCHEMA,
-    SIGNAL_CALLBACK, TYPE)
-from .device import get_mysensors_devices
+    MYSENSORS_GATEWAY_READY, MYSENSORS_GATEWAYS)
+from .handler import HANDLERS
+from .helpers import discover_mysensors_platform, validate_child
 
 _LOGGER = logging.getLogger(__name__)
 
 GATEWAY_READY_TIMEOUT = 15.0
 MQTT_COMPONENT = 'mqtt'
-MYSENSORS_GATEWAY_READY = 'mysensors_gateway_ready_{}'
 
 
 def is_serial_port(value):
@@ -167,23 +163,14 @@ async def _discover_persistent_devices(hass, hass_config, gateway):
     for node_id in gateway.sensors:
         node = gateway.sensors[node_id]
         for child in node.children.values():
-            validated = _validate_child(gateway, node_id, child)
+            validated = validate_child(gateway, node_id, child)
             for platform, dev_ids in validated.items():
                 new_devices[platform].extend(dev_ids)
     for platform, dev_ids in new_devices.items():
-        tasks.append(_discover_mysensors_platform(
+        tasks.append(discover_mysensors_platform(
             hass, hass_config, platform, dev_ids))
     if tasks:
         await asyncio.wait(tasks, loop=hass.loop)
-
-
-@callback
-def _discover_mysensors_platform(hass, hass_config, platform, new_devices):
-    """Discover a MySensors platform."""
-    task = hass.async_create_task(discovery.async_load_platform(
-        hass, platform, DOMAIN,
-        {ATTR_DEVICES: new_devices, CONF_NAME: DOMAIN}, hass_config))
-    return task
 
 
 async def _gw_start(hass, gateway):
@@ -222,112 +209,15 @@ def _gw_callback_factory(hass, hass_config):
     @callback
     def mysensors_callback(msg):
         """Handle messages from a MySensors gateway."""
-        start = timer()
         _LOGGER.debug(
             "Node update: node %s child %s", msg.node_id, msg.child_id)
 
-        _set_gateway_ready(hass, msg)
+        msg_type = msg.gateway.const.MessageType(msg.type)
+        msg_handler = HANDLERS.get(msg_type.name)
 
-        try:
-            child = msg.gateway.sensors[msg.node_id].children[msg.child_id]
-        except KeyError:
-            _LOGGER.debug("Not a child update for node %s", msg.node_id)
+        if msg_handler is None:
             return
 
-        signals = []
+        hass.async_create_task(msg_handler(hass, hass_config, msg))
 
-        # Update all platforms for the device via dispatcher.
-        # Add/update entity if schema validates to true.
-        validated = _validate_child(msg.gateway, msg.node_id, child)
-        for platform, dev_ids in validated.items():
-            devices = get_mysensors_devices(hass, platform)
-            new_dev_ids = []
-            for dev_id in dev_ids:
-                if dev_id in devices:
-                    signals.append(SIGNAL_CALLBACK.format(*dev_id))
-                else:
-                    new_dev_ids.append(dev_id)
-            if new_dev_ids:
-                _discover_mysensors_platform(
-                    hass, hass_config, platform, new_dev_ids)
-        for signal in set(signals):
-            # Only one signal per device is needed.
-            # A device can have multiple platforms, ie multiple schemas.
-            # FOR LATER: Add timer to not signal if another update comes in.
-            async_dispatcher_send(hass, signal)
-        end = timer()
-        if end - start > 0.1:
-            _LOGGER.debug(
-                "Callback for node %s child %s took %.3f seconds",
-                msg.node_id, msg.child_id, end - start)
     return mysensors_callback
-
-
-@callback
-def _set_gateway_ready(hass, msg):
-    """Set asyncio future result if gateway is ready."""
-    if (msg.type != msg.gateway.const.MessageType.internal or
-            msg.sub_type != msg.gateway.const.Internal.I_GATEWAY_READY):
-        return
-    gateway_ready = hass.data.get(MYSENSORS_GATEWAY_READY.format(
-        id(msg.gateway)))
-    if gateway_ready is None or gateway_ready.cancelled():
-        return
-    gateway_ready.set_result(True)
-
-
-def _validate_child(gateway, node_id, child):
-    """Validate that a child has the correct values according to schema.
-
-    Return a dict of platform with a list of device ids for validated devices.
-    """
-    validated = defaultdict(list)
-
-    if not child.values:
-        _LOGGER.debug(
-            "No child values for node %s child %s", node_id, child.id)
-        return validated
-    if gateway.sensors[node_id].sketch_name is None:
-        _LOGGER.debug("Node %s is missing sketch name", node_id)
-        return validated
-    pres = gateway.const.Presentation
-    set_req = gateway.const.SetReq
-    s_name = next(
-        (member.name for member in pres if member.value == child.type), None)
-    if s_name not in MYSENSORS_CONST_SCHEMA:
-        _LOGGER.warning("Child type %s is not supported", s_name)
-        return validated
-    child_schemas = MYSENSORS_CONST_SCHEMA[s_name]
-
-    def msg(name):
-        """Return a message for an invalid schema."""
-        return "{} requires value_type {}".format(
-            pres(child.type).name, set_req[name].name)
-
-    for schema in child_schemas:
-        platform = schema[PLATFORM]
-        v_name = schema[TYPE]
-        value_type = next(
-            (member.value for member in set_req if member.name == v_name),
-            None)
-        if value_type is None:
-            continue
-        _child_schema = child.get_schema(gateway.protocol_version)
-        vol_schema = _child_schema.extend(
-            {vol.Required(set_req[key].value, msg=msg(key)):
-             _child_schema.schema.get(set_req[key].value, val)
-             for key, val in schema.get(SCHEMA, {v_name: cv.string}).items()},
-            extra=vol.ALLOW_EXTRA)
-        try:
-            vol_schema(child.values)
-        except vol.Invalid as exc:
-            level = (logging.WARNING if value_type in child.values
-                     else logging.DEBUG)
-            _LOGGER.log(
-                level,
-                "Invalid values: %s: %s platform: node %s child %s: %s",
-                child.values, platform, node_id, child.id, exc)
-            continue
-        dev_id = id(gateway), node_id, child.id, value_type
-        validated[platform].append(dev_id)
-    return validated
